lab-seguranca/app/requirements.txt
Flask==2.1.1
Werkzeug==2.1.1
psycopg2-binary